style: map legend / toggles
